requirements.txt
pymongo
python-dotenv
requests
pandas
transformers
torch 
torchvision 
torchaudio
sentence-transformers
pinecone-client
numpy
openai
fastapi
uvicorn
streamlit